Lagt til ekstra testing og feilhåndtering tilknyttet sletting av filer

+++ src/main/kotlin/no/nav/soknad/arkivering/soknadsfillager/service/SlettFilerService.kt
@@ -1,20 +1,30 @@
 package no.nav.soknad.arkivering.soknadsfillager.service
 
-import no.nav.soknad.arkivering.soknadsfillager.dto.FilElementDto
 import no.nav.soknad.arkivering.soknadsfillager.repository.FilRepository
+import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
+import kotlin.RuntimeException
 
 @Service
 class SlettFilerService(private val filRepository: FilRepository){
+	private val logger = LoggerFactory.getLogger(javaClass)
 
 	 fun slettFiler (filListe: List<String>){
 
 		 filListe.forEach{slettFil(it)}
-
 	}
 
 	private fun slettFil (uuid: String) {
-
-		filRepository.deleteById(uuid)
+		try {
+			if (filRepository.findById(uuid).isEmpty){
+				logger.error("$uuid er ikke i basen")
+				throw RuntimeException("Fil med $uuid er ikke i basen" )
+			}
+			logger.info("Fil med $uuid er slettet fra basen")
+			filRepository.deleteById(uuid)
+		} catch (e:Exception){
+			logger.error("Feil tilknyttet fil med $uuid ", e)
+			throw e
+		}
 	}
 }
